requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
inflection==0.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
